added APIs of Admin for booking, passenger, payment, user management & partially updated repos by required queries.

+++ SpringBootServer/IndiSky/src/main/java/com/indisky/entities/FlightSeat.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class FlightSeat {
     @Column(name = "seat_id")
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer seatId;
+    private Long seatId;
 
     @ManyToOne
     @JoinColumn(name = "flight_id")
